requirements.txt
# Основные зависимости проекта
numpy>=1.21.0

# Для установки пакета t-tech-investments из частного репозитория T‑Bank
# необходимо указать дополнительный индекс.
# При установке через pip используйте команду:
# pip install -r requirements.txt --extra-index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple
# или установите пакет отдельно:
# pip install t-tech-investments --index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple

--extra-index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple
t-tech-investments>=0.2.0